Modif

+++ backend/controllers/sauces.js
@@ -1,15 +1,18 @@
 const sauces = require("../models/sauces");
-
-exports.createTSauces = (req, res, next) => {
+const fs = require("fs");
+exports.createSauces = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
   delete sauceObject._id;
   delete sauceObject._userId;
   const sauce = new sauces({
     ...sauceObject,
-    userId: req.auth.userId,
     imageUrl: `${req.protocol}://${req.get("host")}/images/${
       req.file.filename
     }`,
+    likes: 0,
+    dislikes: 0,
+    usersLiked: [],
+    usersDisliked: [],
   });
 
   sauce
@@ -37,43 +40,57 @@ exports.getOneSauces = (req, res, next) => {
 };
 
 exports.modifySauces = (req, res, next) => {
-  const sauce = new sauces({
-    _id: req.params.id,
-    title: req.body.title,
-    description: req.body.description,
-    imageUrl: req.body.imageUrl,
-    price: req.body.price,
-    userId: req.body.userId,
-  });
+  const sauceObject = req.file
+    ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+  delete sauceObject._userId;
   sauces
-    .updateOne({ _id: req.params.id }, sauce)
-    .then(() => {
-      res.status(201).json({
-        message: "Sauce modifiée",
-      });
+    .findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        sauces
+          .updateOne(
+            { _id: req.params.id },
+            { ...sauceObject, _id: req.params.id }
+          )
+          .then(() => res.status(200).json({ message: "Objet modifié!" }))
+          .catch((error) => res.status(401).json({ error }));
+      }
     })
     .catch((error) => {
-      res.status(400).json({
-        error: error,
-      });
+      res.status(400).json({ error });
     });
 };
 
 exports.deleteSauces = (req, res, next) => {
   sauces
-    .deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: "Supprimer",
-      });
+    .findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          sauces
+            .deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Objet supprimé !" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
     })
     .catch((error) => {
-      res.status(400).json({
-        error: error,
-      });
+      res.status(500).json({ error });
     });
 };
-
 exports.getAllSauces = (req, res, next) => {
   sauces
     .find()
@@ -85,4 +102,30 @@ exports.getAllSauces = (req, res, next) => {
         error: error,
       });
     });
+};
+
+exports.likeSauces = (req, res, next) => {
+  // Si Like = 1, l'utilisateur like
+  if (req.body.like == 1) {
+    sauces
+      .findOne({ _id: req.params.id })
+      .then((sauce) => {
+        if (!sauce.usersLiked.includes(req.body.userId)) {
+          sauces
+            .updateOne(
+              { _id: req.params.id },
+              {
+                $inc: { likes: 1 },
+                $push: { usersLiked: req.body.userId },
+              }
+            )
+            .then(() => res.status(201).json({ message: "Like +1" }))
+            .catch((error) => res.status(400).json({ error }));
+        }
+      })
+      .catch((error) => res.status(400).json({ error }));
+  }
+  // Si like = 0, l'utilisateur annule son like ou dislike
+
+  // Si like = -1, l'utilisateur dislike
 };
